Update extended XYZ writer to work with GenericFileIO and more
Currently, writing data columns to extended XYZ does not with with
GenericFileIO calculators such as Espresso. This is because the extxyz
writer explicitly checks "isinstance(calculator, Calculator)" but some
perfectly good calculators do not inherit Calculator - they inherit
BaseCalculator.

The check is actually needed to ensure that a .results dict will be
present; this is part of BaseCalculator, so a check against
BaseCalculator instead should be fine.