Swap array indices for get_eigenvalues of GIO / GetOutputsMixin
This array lookup breaks DOS calculations etc. for the Espresso
calculator. Probably it slips through some tests because the (0, 0)
case is not affected. Still it is surprising that it has not yet been
encountered by some other generic calculator; hopefully there is not
some workaround that this will break in turn.