`io.extxyz`: Improve boolean property parsing
Currently, the parsing of boolean properties (e.g. `pbc`) is flawed and
does not respect the extended XYZ specification. For example:

```
pbc="True True False"
```

Will simply be passed as a string to the `pbc` argument of the `Atoms`
constructor. This will in turn interpret this as `True` and hence assing
full periodic boundary conditions, which is not correct.

Here we improve the parsing of booleans in the properties of `extxyz`
files by adapting the `key_val_str_to_dict` function:

1. Add all the allowed strings for booleans to the `str_to_bool`
mapping.
2. In case a string value is found that is not allowed because of the
specification, raise instead of letting it pass silently.