Ignore rsyncdir deprecation warning for full calculator tests
Calculator tests are failing due to deprecations warning. This arises
from incompatibility between pytest-cov and pytest xdist.

Pytest-cov have fixed it in master but who knows when we will have a
new version running on our CI...

This warning is already ignored in other test runs, I assume it was
overlooked for the full Calculators run because that isn't run ever
time.