BUG: Make ODE12r run optimization for steps=1
Suppose the user calls the "run" method (inherited from
ase.optimize.sciopt.SciPyOptimize) of an instance of the ODE12r class
(or a subclass who also calls ase.optimize.ode.ode12r) with steps=0. In,
the previous implementation, no optimization is run. This is an expected
and intuitive behaviour that is consistent with other ASE-defined
subclasses of ase.optimize.Optimizer. However, no optimization is run if
the "run" method is called with steps=1 either. This fix leads to more
intuitive behaviour of the ODE12r optimizer.

Closes: #1267