Add more descriptive type hints
As shown with the newly added tests, the addition of more descriptive type
hints to the `ase.utils.deprecated` decorator serves as an additional
safeguard against developers accidentally defining a callback function that
modifies the arguments of the deprecated function without explicitly returning
a value. If silently modifying function arguments is desired, the type error
can be suppressed by an inline `type: ignore[arg-type, return-value]` comment.